Charlie's Work/requirements.txt
# AI Screening Interview - Dependencies
# =====================================

# Core LLM (choose one or both)
openai>=1.50.0            # For OpenAI (GPT-4) - also used for voice
# anthropic>=0.40.0       # Optional: For Claude (uncomment if preferred)

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Web Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
redis>=5.0.0
python-multipart>=0.0.9

# Speech (for voice interview mode)
# deepgram-sdk>=3.0.0     # Optional: for real-time STT
# elevenlabs>=1.0.0       # Optional: for natural TTS

# Audio Processing (for voice interface)
sounddevice>=0.5.0
numpy>=2.0.0
pydub>=0.25.0             # For audio format conversion

# Development
pydantic>=2.9.0
python-dotenv>=1.0.0
